#ServiceInfo доработка логики если заказ уже был создан.

+++ network/src/main/java/com/photoprint/network/api/model/order/Order.java
@@ -8,9 +8,9 @@ import com.google.gson.annotations.SerializedName;
 public class Order {
 
     @SerializedName("order_id")
-    private Integer orderId;
+    private long orderId;
 
-    public Integer getOrderId() {
+    public long getOrderId() {
         return orderId;
     }
